+++ apps/server/src/main/java/com/weanet/server/api/dashboard/service/DashboardService.java
@@ -3,8 +3,8 @@ package com.weanet.server.api.dashboard.service;
 import com.weanet.server.api.report.dto.ReportResponse;
 import com.weanet.server.api.report.repository.ReportRepository;
 import com.weanet.server.api.route.domain.Route;
-import com.weanet.server.api.route.dto.RouteIntegratedReportResponse;
-import com.weanet.server.api.route.dto.RouteStepResponse;
+import com.weanet.server.api.route.dto.request.*;
+import com.weanet.server.api.route.dto.response.*;
 import com.weanet.server.api.route.repository.RouteRepository;
 import com.weanet.server.api.route.service.RouteAdviceService;
 import com.weanet.server.api.route.service.RouteEnrichmentService;
