+++ apps/server/src/main/java/com/weanet/server/api/route/dto/response/RouteDetailResponse.java
@@ -1,4 +1,4 @@
-package com.weanet.server.api.route.dto;
+package com.weanet.server.api.route.dto.response;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AllArgsConstructor;
