+++ apps/server/src/main/java/com/weanet/server/api/route/dto/response/RouteSearchStepResponse.java
@@ -1,4 +1,4 @@
-package com.weanet.server.api.route.dto;
+package com.weanet.server.api.route.dto.response;
 
 import com.weanet.server.api.route.domain.TransportType;
 import io.swagger.v3.oas.annotations.media.Schema;
