+++ apps/server/src/main/java/com/weanet/server/api/route/service/RouteAdviceService.java
@@ -1,7 +1,9 @@
 package com.weanet.server.api.route.service;
 
-import com.weanet.server.api.route.dto.RouteIntegratedReportResponse;
-import com.weanet.server.api.route.dto.RouteStepResponse;
+import com.weanet.server.api.route.dto.request.*;
+import com.weanet.server.api.route.dto.response.*;
+import com.weanet.server.api.route.dto.request.*;
+import com.weanet.server.api.route.dto.response.*;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
