+++ apps/server/src/main/java/com/weanet/server/api/route/service/RouteEnrichmentService.java
@@ -4,7 +4,8 @@ import com.weanet.server.api.external.service.CongestionService;
 import com.weanet.server.api.external.service.SubwayService;
 import com.weanet.server.api.external.util.KmaCoordinateConverter;
 import com.weanet.server.api.route.domain.TransportType;
-import com.weanet.server.api.route.dto.RouteStepResponse;
+import com.weanet.server.api.route.dto.request.*;
+import com.weanet.server.api.route.dto.response.*;
 import com.weanet.server.api.weather.dto.WeatherResponse;
 import com.weanet.server.api.weather.service.WeatherService;
 import com.weanet.server.api.external.dto.SubwayRealtimeResponse;
