+++ apps/server/src/main/java/com/weanet/server/api/route/service/RouteService.java
@@ -2,7 +2,8 @@ package com.weanet.server.api.route.service;
 
 import com.weanet.server.api.external.service.ExternalMapService;
 import com.weanet.server.api.route.domain.Route;
-import com.weanet.server.api.route.dto.*;
+import com.weanet.server.api.route.dto.request.*;
+import com.weanet.server.api.route.dto.response.*;
 import com.weanet.server.api.route.repository.RouteRepository;
 import com.weanet.server.global.error.exception.BusinessException;
 import com.weanet.server.global.error.exception.ErrorCode;
